+++ site/src/theme/MDXComponents.tsx
@@ -1,7 +1,0 @@
-import MDXComponents from '@theme-original/MDXComponents';
-import InteractiveTutorial from '@site/src/components/InteractiveTutorial';
-
-export default {
-	...MDXComponents,
-	InteractiveTutorial,
-};