+++ tutorials/_site/src/theme/MDXComponents.tsx
@@ -1,0 +1,9 @@
+import MDXComponents from '@theme-original/MDXComponents';
+import InteractiveTutorial from '../components/InteractiveTutorial';
+
+export default {
+	...MDXComponents,
+	InteractiveTutorial,
+} satisfies typeof MDXComponents & {
+	InteractiveTutorial: typeof InteractiveTutorial;
+};